requirements.txt
pandas>=2.0.0
numpy>=1.24.0
lightgbm>=4.0.0
gplearn>=0.4.2
baostock>=0.8.8
pyyaml>=6.0
tushare>=1.2.89
scikit-learn>=1.3.0
matplotlib>=3.7.0
fire>=0.5.0
